Replaced some sounds with unclear copyright with new ones from 0.3.x and changed frequencies all to 32000hz
Changed the files to 32000hz to fix the sound pitch bug and to avoid having to resample
to 32000hz.  The space gained from compressing the images in the last commit should
make up for the extra space requirements for this and there shouldn't be any noticeable
difference in frame output rate.